requirements.txt
flask==3.1.3
werkzeug==3.1.7
jinja2==3.1.6
markupsafe==3.0.3
click==8.3.1
itsdangerous==2.2.0
python-dotenv==1.2.1
google-genai==1.68.0
gunicorn==23.0.0
